Fixed overflow_fix issue

+++ tests/openvino/native/models.py
@@ -512,7 +512,7 @@ class SharedConvModel(OVReferenceModel):
         input_1 = opset.parameter(input_shape, name=input_name)
         if kernel is None:
             c_in = input_shape[1]
-            kernel = self._rng.random((3, c_in, 1, 1)).astype(np.float32)
+            kernel = self._rng.random((3, c_in, 1, 1))
         const_kernel = opset.constant(kernel, np.float32, name='Shared_conv_w')
         strides = [1, 1]
         pads = [0, 0]
